Add main.go
